Modified scheming_multiple_choice validator
Previously, if choice_values list was empty and field values existed,
validation would fail.

Now, if choice_values is empty and values exist, those values are
presumed to be valid.

This fix is being made to address the circular logic case where a
choice_list is temporarily unavailable due to resource update order when
datapusher-plus is doing a resubmit.